Temporarily remove `tracestate`
The usage of the W3C `tracestate` for propagating OpenTelemetry-specific
data is currently ambiguous, and depends on some as yet unmade decisions
about how distributed context propagation.

In order to prevent breaking changes in the near-ish future, it feels
safest to temporarily remove `tracestate` from the spec so that its use
can be additive once these questions are resolved.