New spanids should always be created for server spans.
Fixes #359. Standardizing the behavior of the creation of server span
ids helps ensure server behavior will be the same regardless of the
language used.

In practice, the negative consequences of SpanID being different for
client and server in storage systems that expect it are minimal,
at worst appearing as separate spans.