Producer SpanKind is not always a remote call
It's popular in asynchronous scenarios to send messages in batches.

In this case, each message may (and should) carry a unique context to be individually traceable. Arguably it means creating span per message.

Sending a batch (which is remote outgoing call) is represented by a different span that is linked to all of the messages span contexts.

So async messaging scenarios with batching involve
* *message* spans and they map to PRODUCER kind (they are parent of asynchronous request)
* *send* span that is CLIENT kind (which is remote outgoing call). But it is not a parent of any asynchronous request

In some cases (not involving batching) it is possible that dedicated *message* span is not needed and *send* may represent both: context in the message and remote outgoing call. In this case, such span should have PRODUCER kind