Propose OpenTelemetry Logs Vision (#92)
* Propose OpenTelemetry Logs Vision

This change proposes high-level items that define our long-term vision for
Logs support in OpenTelemetry project, what we aspire to achieve.

This is a starting point for discussion. I believe it is important for us
to collectively come up with such vision.

Comments, thoughts and opinions are highly welcome!
Propose OpenTelemetry Logs Vision

This change proposes igh-level items that define our long-term vision for
Logs support in OpenTelemetry project, what we aspire to achieve.

This a vision document that reflects our current desires. It is not a commitment
to implement everything precisely as listed. The primary purpose of this
document is to ensure all contributors work in alignment. As our vision changes
over time maintainers reserve the right to add, modify, and remove items from
this document.

* Address PR comments