Adds usage explanation from Zipkin
People used to cause a lot of problems by adding UUID in service for lack of any explanation of impact.  I've copied some notes from our docs.

I would like to mention that this will be very problematic to allow this to have a scope besides local root. I'm not sure the local root concept exists here, or what scope resource is allowed to be in. If Resource is required to be a scope larger than local root, should be ok.