Define Log Data model (#97)
* Define Log Data model

This is a proposal of a data model and semantic conventions that allow to
represent logs from various sources: application log files, machine
generated events, system logs, etc.

Existing log formats can be unambiguously mapped to this data model.
Reverse mapping from this data model is also possible to the extent that
the target log format has equivalent capabilities.

The purpose of the data model is to have a common understanding of what a
log record is, what data needs to be recorded, transferred, stored and
interpreted by a logging system.

* Move content from Google Doc to markdown here

* Address PR comments

* Add Google Cloud Logging mapping

* Address PR comments

* Resolve Open Questions

* Add ECS mapping

* Address PR comments