Add logs overview
The overview.md is the summarized version of ideas that were discussed
multiple times in meetings or in Gitter channels.

Note that the document is written as if we already have implemented
all desirable functionality, thus there are multiple "link TBD" entries
in the document. We will add the links as the implementations are done.