Update SignalFx Event/Log conventions (#127)
Per my conversation with @tigrannajaryan, we have determined that event dimensions are
more appropriate as attributes since a single logical Resource could emit events with
disparate dimension sets.  Also, we need to be able to unambiguously reverse a LogRecord to
a SignalFx Event, so properties are nested in a specially keyed map inside Attributes.