Add specification/SDK compliance matrix
This commit introduces the compliance matrix. It is not yet a complete
list of features that we want to track. I marked several sections as
TBD. We will need to add features/rows in those sections in separate PRs.
In the meantime we can ask SIGs to add +, - or N/A in the sections
that are more or less ready (currently Traces API and OTLP, Zipkin Exporters).

Contributes to https://github.com/open-telemetry/opentelemetry-specification/issues/827