Shorten otel.instrumentation_library.* attribute names (#989)
otel.instrumentation_library.name and otel.instrumentation_library.version
are quite long. The keys increase the size of the wire data (and they are present
for every span emitted by OpenTelemetry).

Long keys may also cause trouble when formatting in the UI.

This change uses a shorter "otel.library" prefix. I think this does not create
any ambiguity, it appears that this is the only "library" OpenTelemetry-emitted
span should refer to.