Change default OTLP port number
Contributes to https://github.com/open-telemetry/opentelemetry-specification/issues/1148

Note that a separate port is used for OTLP/HTTP for now. There is currently work
in progress to confirm that we can use the same port. Once we have the confirmation
I will update the spec again to use one port.