Clarify when OpenTelemetry MUST provide a context implementation
Language implementations should review the OpenTelemetry context if there is an established context implementation in their ecosystem to see if OpenTelemetry requirements are satified by the existing context. If not, OpenTelemetry must provide a Context implementation.

I'm moving the language-specific note to the top and calling out that Context is an in-process context propagation mechanism.