Rename "faas" to function
The resource semantic conventions for serverless function uses "faas"
as a prefix. FaaS would be a name to identify the concept of "function
as a service", not the platforms or the workloads.
The semantic conventions we define here are used to
attribute to the functions not to the concept.
FaaS is also not as clear as "function" to developers who are not
familiar with industry terms such as IaaS/PaaS/FaaS.

This change renames FaaS to function to better represent that
the resource was a serverless function.