Clarify that Events and Log Records are conceptually the same thing (#1461)
I get this question a lot. This clarifies that events and log records
are different names of the same concept (from Otel perspective).

This was discussed verbally many times and I believe it was also
in my initial logs proposal but got lost somewhere in the process.

This also fixes the TBD link in the spec that was supposed to point
to this explanation.