Refine log collection story based on Stanza contribution
Due to the recent contribution of Stanza log collection library
we have a updated understanding of how log collection can work.

We now believe that in certain cases the Colector can do this work
successfully.

This PR updates the relevant parts of the spec and includes links
to some recently developed Collector components.