Scope baggage/context interaction functionality
The Baggage API states that it can be implemented by providing
functionality interacting solely with the `Context`. If an
implementation chooses this option it follows that it should not be
required to provide access to a Baggage object. This updates the Context
Interaction section of the Baggage API to clarify this caveat.