Avoid using "infrastructure" terminology for all cloud
Infrastructure in cloud has a specific connotation for concepts
explained in Infrastructure-as-a-Service (IaaS). Avoid this
termilogy because it's exclusive of some other services/platforms
that is represented by the "cloud" namespace.

Renaming "infrastructure_service" to "service" also due to the new
key length limits we are considering and infrastructure_service is
a lengthy choice.