Clarify the TraceIdRatioBased description format
Samplers need to return a static string as a description. The string the
TraceIdRatioBased sampler returns needs to be based on the sampling
ratio that it is configured for. The current specification could be
misinterpreted as requiring a static string that was not based on the
samplers sampling ratio be used. This corrects that discrepancy.