Rename DRAGONFLYBSD to DRAGONFLY
The `os.type` values seem inspired by the GOOS list, except for the
one for Dragonfly BSD, which differs from the GOOS list.

Renaming the constant to DRAGONFLY makes complying with the spec on
Go implementations easier, removing the need to have a special case
for this operating system family.