Clarify "key/value pair list" vs "map" in Log Data Model
Parts of the data model documents were using the term "map", the others
were using the term "key/value pair list" without clearly telling why
and how they are the same.

This change clarifies and ensures that we refer to a map of key/value pairs
that can be represented in different way in different languages.

Resolves https://github.com/open-telemetry/opentelemetry-specification/issues/1592