Add schema files and a script with minimal checks that verifies them
This adds some minimal sanity checks to make sure we don't forget
to add the schema file when we make a release. The checks are tied
to the CHANGELOG.md.

In the future we can do more thorough verifications, e.g. parse the content
of the schema YAML file and ensure it includes the changes that we can
detect by looking at the changes at the semantic convention files.