clarify meaning of "Certificate File"
The current version of the specification makes me think that Certificate File is intended to be used as part of the client authentication, which it is not. This change addresses part of the confusion discussed in https://github.com/open-telemetry/opentelemetry-specification/issues/1363, the certificate file option and environment variable points to the certificate used to verify the server's certificate. 

See also https://github.com/open-telemetry/opentelemetry-specification/pull/1375.