Add automated Schema File publish action
The action should be triggered manually after every spec release
and will copy/update schema files from this repository to the
website repository so that they become available publicly.

The publish-schemas action after manual triggering will create a PR
in https://github.com/open-telemetry/opentelemetry.io repo, and the
PR needs to be approved and merged manually.

This is similar to how we publish doc updates from language repos
to the website repo.

Resolves https://github.com/open-telemetry/opentelemetry.io/issues/577
Resolves https://github.com/open-telemetry/opentelemetry-specification/issues/1693