Add missing apply_to_spans section
apply_to_spans section was missing although it is implied to
be present in the description of how span changes section
is supposed to work.

This fixes the omission by adding apply_to_spans section.
This is not a new feature, there is no semantics change, just
fixing a bug in file format description.