Simplify Baggage handling in the OpenTracing Shim layer.
Baggage is NOT directly associated with an OpenTelemetry Span
directly anymore, but with the OpenTracing Shim, which
greatly simplifies its handling.

This comes with the recommendation of NOT consuming the
OpenTelemetry API and the OpenTracing Shim the codebase,
as the Span's associated Baggage may be lost during
in-process or inter-process Propagation.