Clarify the span Temporal relationship diagram (#2204)
* Clarify the span Temporal relationship diagram

The temporal span relationship diagram on the overview page can be a little misleading at first glance. Due to Span C being both shorter and below Span B in the diagram a first reading suggests that it is a sub span of B. When compared to the tree view you can sort out that it's not meant to be, but it adds a bit of cognitive friction that's unnecessary. To more quickly communicate intent, I've shorten Spans B and D, and pushed Span C out past B.

* Update specification/overview.md to show children spans outliving parents.

Co-authored-by: Armin Ruech <armin.ruech@dynatrace.com>

Co-authored-by: Armin Ruech <armin.ruech@dynatrace.com>
Co-authored-by: Joshua MacDonald <jmacd@users.noreply.github.com>