Add `make all` target to test changes locally
Make all performs roughly the same steps as the alrady existing github actions. The only
exception is yamllint which I was not able to make work on my Mac after quite some time
and had to skip. If anyone knows how to make it work on a Mac please submit a follow-up PR.

Note that the checks can run in parallel to speed up, e.g. `make -j4`.