Declare Log Data Model Stable
This is the first call to everyone that we are planning to declare
log data model stable.

All issues that were identified to be required for this declaration
are closed: https://github.com/open-telemetry/opentelemetry-specification/issues?q=is%3Aopen+is%3Aissue+label%3Arelease%3Arequired-logdatamodel-ga

Anyone who believes there are reasons NOT to declare the data model
stable please speak now.

We need to keep this PR open for a while and circulate this intent
widely.