Clarify that issues must be created before PRs
The contributing guide says all significant changes should go through OTEP
process. It also describes an issue workflow which says that an issue is
created first, the issue is triaged and accepted and then a PR is created,
however the "Proposing a change" section does not explicitly describe this
as a possibility.

I added a paragraph to make it clear that there are 3 possible options:
with an OTEP, with an issue and trivial changes that are just PRs.

Unless we make this change it appears that everything that is not a trivial
change must go through OTEP, which is not our current process and is not
what we want.