Clarify that Scope is defined at build time
I have seen several time a confusion around the nature of the
scope and whether scope attributes can change at runtime.

The purpose of this PR is to mak sure we all agree the scope is
a build time concept, or if we disagree then explicitly specify
what else the scope can denote and how its attribute can change
at runtime.