Clarification on what an Event is, and what the event.domain and event.name attributes represent (#2848)
* Remove language around uniqueness of event names within a domain

This came up in
https://github.com/open-telemetry/opentelemetry-collector-contrib/issues/14474
where it was highlighted that just `name` is not sufficient to provide
for uniqueness of a Kubernetes Event.

The attribute `event.domain` is only meant to be a logical separation
across different event systems, so it allows for similar looking names
in different domains but with a completely different purpose. However,
within each domain we don't have to have any uniqueness constraints on
the `name` and leave it to whatever is idiomatic to the particular event
system the domain represents.