Make sure it is very clear we are not building a Logging API
We are only building a logging "backend" which is supposed to be used via appenders by existing logging libraries. Additionally, we may provide helpers to build "events" that make easier to build LogRecords with specific semantic convention. Such helpers will also call the logging "backend" APIs.

We do not intend to have user-facing logging API in the spec.

Resolves https://github.com/open-telemetry/opentelemetry-specification/issues/2936