Suggest column # as an extra source code attribute
In some languages such as Javascript or Haskell, it's common to have multiple lambdas/anonymous functions defined on the same line. Without a column number, it's impossible to accurately pinpoint which of the anonymous functions in the same file and the same line we are talking about. A column attribute is necessary to fully specify which unit of code we are interested in.