Fix Hist param, value instead of `Measurement`
A `Measurement` is defined as the encapsulation of a value and
attributes. The Histogram record API currently is defined to accept a
`Measurement` and attributes. This means, if implemented as currently
defined, a user would pass a value/attributes pair and then additional
attributes. It is not likely this was the original intent of the
specified parameters.