Remove No-Op instrument and Meter creation restrictions (#3322)
The current No-Op requires all instruments to be created from a Meter
and a Meter to be created from a MeterProvider. This was inherited from
the SDK specification, but it is not necessary. The No-Op performs no
operations and hold no state, therefore creating these objects directly
should not have logical or processing consequences. If there are
consequences of doing this in a language implementation, they are still
permitted to keep this restriction.

This change is desired by the Go SIG. We want to export these No-Op
types so users can create them directly. This will allow them to embed
these No-Op types in their SDKs to define default behaviour, and allow
them to use them directly for testing (instead of having to go through
the creation chain all starting with a MeterProvider).

---------

Co-authored-by: Reiley Yang <reyang@microsoft.com>
Co-authored-by: Joshua MacDonald <jmacd@users.noreply.github.com>