Clarify "Emit a LogRecord" API
We discussed in Log SIG today and the current was unclear since
it could be interpreted as API taking LogRecord as an input.
This is not the case. The effect of making this API call is
to emit a LogRecord to the pipeline, but the API does not
take a LogRecord as an input. The input parameters are described
in a subsequent paragraph.