Specify MeterProvider configurable cardinality limits (#2960)
Fixes #1891.

**EDIT: Updated to specify cardinality limits at the View/Instrument
level with a Reader-level default. Updated to use a hard limit**

## Changes

Adds optional support for a maximum cardinality limit.

The recommended default is 2000, based on this comment by
https://github.com/open-telemetry/opentelemetry-specification/issues/1891#issuecomment-967672167
@jack-berg.

~The Prometheus-WG SIG discussed this on Nov 9, 2022 and reached this
recommended solution to the problem outlined in #1891. The consequence
of exceeding these limits is in line with the current Prometheus server
behavior, which drops targets that misbehave. The discussed was
summarized here:
https://github.com/open-telemetry/opentelemetry-specification/issues/1891#issuecomment-1309089908~