Clarify that schema transformations SHOULD overwrite input data
Resolves: https://github.com/open-telemetry/opentelemetry-specification/issues/3497

Alternates Considered
=====================

Instead of requiring overwriting the input data we could say that the
transformation SHOULD be aborted if a conflict is detected or that transformation
SHOULD be ignored if a conflict is detected.

The chosen approach (overwriting) seems to be best balance between maintaining
reasonable outcome (keep the more important data, discarding the less important data)
while also having minimal complexity of implementation (aborting and reporting errors
is more complicated and more costly).

Why "SHOULD" and not "MUST"?
============================

There are significant tradeoffs involved in what happens in the schema transform
and we believe there may be valid use cases where after careful consideration a
different approach may be taken (e.g. reject and surface the problem to the end user).
We do not want to prohibit these other approaches. The goal of this change is to merely
have a reasonable default interpretation of the schema transformation rules.