Clarify identical instrument definition for SDK.
Currently, the SDK states that identical instruments need to have their
data aggregated together. It is implied that the definition of
"identical" is defined by the metric data-model. However, the linked
data-model defines identical in a way that does not include the
Instrument number type. Therefore, if two instruments are "identical" in
all their fields, but one if defined for double-precision floating point
values and the other integer values they will need to be aggregated
together. This is challenging if not impossible for strongly typed
implementations.

Instead, have the SDK explicitly use the definition added to the metric
API defining the number type (and all other Language-level features) as
identifying and therefore included in the definition of identical.