Revise the exemplar default reservoirs
Make it normative that an SDK is recommended to include both types of
default reservoirs.

Make the description of both default reservoirs their own sub-sections.

Correct the "SimpleExemplarReservoir" section to be titled
"SimpleFixedSizeExemplarReservoir".

Loosen the SimpleFixedSizeExemplarReservoir to allow any
uniformly-weighted sampling algorithm instead of just the naive one that
is not optimal.