Recommend a default size of 1 for the SimpleFixedSizeExemplarReservoir
This reservoir type is used for all aggregations other than a histogram
with more than one bucket. Each attribute set the aggregation records
will have reservoir. Therefore, limiting this to a small value by
default when enabled is preferable.

This does not address the way a user will configure this value. That is
left for a future PR/Issue.

Part of #2421