Fix description of OTEL_ATTRIBUTE_COUNT_LIMIT
The OTEL_ATTRIBUTE_COUNT_LIMIT variable was described in the
SDK environment variables document as the "maximum allowed _span_
attribute count" (emphasis added). The "span" in that sentence was
likely a mistake, since this variable applies to attributes in
general, and not only to span attributes. These are more appropriately
controlled with the OTEL_SPAN_ATTRIBUTE_COUNT_LIMIT, which is probably
where this description was accidentally copied from.