Add upgrading and version management documentation (#3695)
Long ago, we defined upgrading procedures for OpenTelemetry, in [OTEP
147](https://github.com/open-telemetry/oteps/blob/main/text/0147-upgrade-procedures.md).
However, we never added this language to the specification.

This pull request adds the OTEP to the specification, verbatim, except
for an addition at the top which clearly spells out dependency
management for application developers, library maintainers, and SDK
implementors.

It has been two years, so I am open to changes to this document. Both
for clarity, and to conform to how we currently practice upgrading and
dependency management.